Disable servo outputs until their positions are set
Upon power-up each servo output previously initially output
middle-of-the-range pulse.  A team requested that this was changed.

This patch adds a new 'enabled' array that stores whether each servo
is enabled or not.  This initially is filled with 'false', and when a
servo is set its corresponding entry is set to 'true'.

In order to reduce complexity, the timer still runs for disabled
servos, but no output will be produced during this time.